stacks/titanic_model/deploy/requirements.txt
scikit-learn==1.5.2 
db-dtypes==1.3.0 
pandas==2.2.3 
google-cloud-storage==2.18.2
Flask==3.0.3